Create Count_con_by_host
A one line style command to get all connected hosts on specific port. As an example here I used the SQL port. It will also try to get the host name from /etc/hosts and get you the total of connections.
Useful to get a quick idea of connected hosts